+++ signalwire/relay_calling.go
@@ -8,7 +8,7 @@ import (
 )
 
 // RelayPhoneDial make outbound phone call
-func (relay *RelaySession) RelayPhoneDial(ctx context.Context, call *CallSession, fromNumber string, toNumber string, timeout uint) error {
+func (relay *RelaySession) RelayPhoneDial(ctx context.Context, call *CallSession, fromNumber string, toNumber string, timeout uint, payload **json.RawMessage) error {
 	var err error
 
 	if relay == nil {
@@ -77,7 +77,7 @@ func (relay *RelaySession) RelayPhoneDial(ctx context.Context, call *CallSession
 }
 
 // RelayPhoneConnect TODO DESCRIPTION
-func (relay *RelaySession) RelayPhoneConnect(ctx context.Context, call *CallSession, fromNumber string, toNumber string) error {
+func (relay *RelaySession) RelayPhoneConnect(ctx context.Context, call *CallSession, fromNumber string, toNumber string, payload **json.RawMessage) error {
 	if relay == nil {
 		return errors.New("empty relay object")
 	}
@@ -141,7 +141,7 @@ func (relay *RelaySession) RelayPhoneConnect(ctx context.Context, call *CallSess
 }
 
 // RelayConnect TODO DESCRIPTION
-func (relay *RelaySession) RelayConnect(ctx context.Context, call *CallSession, ringback *[]RingbackStruct, devices *[][]DeviceStruct) error {
+func (relay *RelaySession) RelayConnect(ctx context.Context, call *CallSession, ringback *[]RingbackStruct, devices *[][]DeviceStruct, payload **json.RawMessage) error {
 	if relay == nil {
 		return errors.New("empty relay object")
 	}
@@ -193,7 +193,7 @@ func (relay *RelaySession) RelayConnect(ctx context.Context, call *CallSession, 
 }
 
 // RelayCallAnswer TODO DESCRIPTION
-func (relay *RelaySession) RelayCallAnswer(ctx context.Context, call *CallSession) error {
+func (relay *RelaySession) RelayCallAnswer(ctx context.Context, call *CallSession, payload **json.RawMessage) error {
 	if len(call.CallID) == 0 {
 		Log.Error("no CallID\n")
 
@@ -231,7 +231,7 @@ func (relay *RelaySession) RelayCallAnswer(ctx context.Context, call *CallSessio
 }
 
 // RelayCallEnd TODO DESCRIPTION
-func (relay *RelaySession) RelayCallEnd(ctx context.Context, call *CallSession) error {
+func (relay *RelaySession) RelayCallEnd(ctx context.Context, call *CallSession, payload **json.RawMessage) error {
 	if len(call.CallID) == 0 {
 		Log.Error("no CallID\n")
 
@@ -285,7 +285,7 @@ func (relay *RelaySession) RelayOnInboundAnswer(ctx context.Context) (*CallSessi
 	}
 
 	if call != nil {
-		err := relay.RelayCallAnswer(ctx, call)
+		err := relay.RelayCallAnswer(ctx, call, nil)
 		if err != nil {
 			return nil, err
 		}
@@ -295,7 +295,7 @@ func (relay *RelaySession) RelayOnInboundAnswer(ctx context.Context) (*CallSessi
 }
 
 // RelayPlayAudio TODO DESCRIPTION
-func (relay *RelaySession) RelayPlayAudio(ctx context.Context, call *CallSession, ctrlID string, url string) error {
+func (relay *RelaySession) RelayPlayAudio(ctx context.Context, call *CallSession, ctrlID string, url string, payload **json.RawMessage) error {
 	if len(call.CallID) == 0 {
 		Log.Error("no CallID\n")
 
@@ -311,11 +311,11 @@ func (relay *RelaySession) RelayPlayAudio(ctx context.Context, call *CallSession
 		Params: playAudioParams,
 	}}
 
-	return relay.RelayPlay(ctx, call, ctrlID, play)
+	return relay.RelayPlay(ctx, call, ctrlID, play, payload)
 }
 
 // RelayPlayTTS TODO DESCRIPTION
-func (relay *RelaySession) RelayPlayTTS(ctx context.Context, call *CallSession, ctrlID string, text, language, gender string) error {
+func (relay *RelaySession) RelayPlayTTS(ctx context.Context, call *CallSession, ctrlID string, text, language, gender string, payload **json.RawMessage) error {
 	if len(call.CallID) == 0 {
 		Log.Error("no CallID\n")
 
@@ -333,11 +333,11 @@ func (relay *RelaySession) RelayPlayTTS(ctx context.Context, call *CallSession, 
 		Params: playTTSParams,
 	}}
 
-	return relay.RelayPlay(ctx, call, ctrlID, play)
+	return relay.RelayPlay(ctx, call, ctrlID, play, payload)
 }
 
 // RelayPlayRingtone TODO DESCRIPTION
-func (relay *RelaySession) RelayPlayRingtone(ctx context.Context, call *CallSession, ctrlID string, name string, duration float64) error {
+func (relay *RelaySession) RelayPlayRingtone(ctx context.Context, call *CallSession, ctrlID string, name string, duration float64, payload **json.RawMessage) error {
 	if len(call.CallID) == 0 {
 		Log.Error("no CallID\n")
 
@@ -354,11 +354,11 @@ func (relay *RelaySession) RelayPlayRingtone(ctx context.Context, call *CallSess
 		Params: playRingtoneParams,
 	}}
 
-	return relay.RelayPlay(ctx, call, ctrlID, play)
+	return relay.RelayPlay(ctx, call, ctrlID, play, payload)
 }
 
 // RelayPlaySilence TODO DESCRIPTION
-func (relay *RelaySession) RelayPlaySilence(ctx context.Context, call *CallSession, ctrlID string, duration float64) error {
+func (relay *RelaySession) RelayPlaySilence(ctx context.Context, call *CallSession, ctrlID string, duration float64, payload **json.RawMessage) error {
 	if len(call.CallID) == 0 {
 		Log.Error("no CallID\n")
 
@@ -374,11 +374,11 @@ func (relay *RelaySession) RelayPlaySilence(ctx context.Context, call *CallSessi
 		Params: playSilenceParams,
 	}}
 
-	return relay.RelayPlay(ctx, call, ctrlID, play)
+	return relay.RelayPlay(ctx, call, ctrlID, play, payload)
 }
 
 // RelayPlay TODO DESCRIPTION
-func (relay *RelaySession) RelayPlay(ctx context.Context, call *CallSession, controlID string, play []PlayStruct) error {
+func (relay *RelaySession) RelayPlay(ctx context.Context, call *CallSession, controlID string, play []PlayStruct, payload **json.RawMessage) error {
 	if len(call.CallID) == 0 {
 		Log.Error("no CallID\n")
 
@@ -395,6 +395,9 @@ func (relay *RelaySession) RelayPlay(ctx context.Context, call *CallSession, con
 			Play:      play,
 		},
 	}
+
+	/* prepare payload per Action in case user want to inspect it. We'll not send this, jsonrpc2 lib will do it's own marshaling on v */
+	savePayload(payload, v)
 
 	call.Lock()
 
@@ -415,7 +418,7 @@ func (relay *RelaySession) RelayPlay(ctx context.Context, call *CallSession, con
 
 	var ReplyBladeExecuteDecode ReplyBladeExecute
 
-	reply, err := relay.Blade.BladeExecute(ctx, &v, &ReplyBladeExecuteDecode)
+	reply, err := relay.Blade.BladeExecute(ctx, &v, &ReplyBladeExecuteDecode /*, payload*/)
 	if err != nil {
 		return err
 	}
@@ -435,7 +438,7 @@ func (relay *RelaySession) RelayPlay(ctx context.Context, call *CallSession, con
 }
 
 // RelayPlayVolume TODO DESCRIPTION
-func (relay *RelaySession) RelayPlayVolume(ctx context.Context, call *CallSession, ctrlID *string, vol float64) error {
+func (relay *RelaySession) RelayPlayVolume(ctx context.Context, call *CallSession, ctrlID *string, vol float64, payload **json.RawMessage) error {
 	if len(call.CallID) == 0 {
 		Log.Error("no CallID\n")
 
@@ -452,6 +455,8 @@ func (relay *RelaySession) RelayPlayVolume(ctx context.Context, call *CallSessio
 			Volume:    vol,
 		},
 	}
+
+	savePayload(payload, v)
 
 	var ReplyBladeExecuteDecode ReplyBladeExecute
 
@@ -475,7 +480,7 @@ func (relay *RelaySession) RelayPlayVolume(ctx context.Context, call *CallSessio
 }
 
 // RelayPlayResume TODO DESCRIPTION
-func (relay *RelaySession) RelayPlayResume(ctx context.Context, call *CallSession, ctrlID *string) error {
+func (relay *RelaySession) RelayPlayResume(ctx context.Context, call *CallSession, ctrlID *string, payload **json.RawMessage) error {
 	if len(call.CallID) == 0 {
 		Log.Error("no CallID\n")
 
@@ -491,6 +496,8 @@ func (relay *RelaySession) RelayPlayResume(ctx context.Context, call *CallSessio
 			ControlID: *ctrlID,
 		},
 	}
+
+	savePayload(payload, v)
 
 	var ReplyBladeExecuteDecode ReplyBladeExecute
 
@@ -514,7 +521,7 @@ func (relay *RelaySession) RelayPlayResume(ctx context.Context, call *CallSessio
 }
 
 // RelayPlayPause TODO DESCRIPTION
-func (relay *RelaySession) RelayPlayPause(ctx context.Context, call *CallSession, ctrlID *string) error {
+func (relay *RelaySession) RelayPlayPause(ctx context.Context, call *CallSession, ctrlID *string, payload **json.RawMessage) error {
 	if len(call.CallID) == 0 {
 		Log.Error("no CallID\n")
 
@@ -530,6 +537,8 @@ func (relay *RelaySession) RelayPlayPause(ctx context.Context, call *CallSession
 			ControlID: *ctrlID,
 		},
 	}
+
+	savePayload(payload, v)
 
 	var ReplyBladeExecuteDecode ReplyBladeExecute
 
@@ -553,7 +562,7 @@ func (relay *RelaySession) RelayPlayPause(ctx context.Context, call *CallSession
 }
 
 // RelayPlayStop TODO DESCRIPTION
-func (relay *RelaySession) RelayPlayStop(ctx context.Context, call *CallSession, ctrlID *string) error {
+func (relay *RelaySession) RelayPlayStop(ctx context.Context, call *CallSession, ctrlID *string, payload **json.RawMessage) error {
 	if len(call.CallID) == 0 {
 		Log.Error("no CallID\n")
 
@@ -569,6 +578,8 @@ func (relay *RelaySession) RelayPlayStop(ctx context.Context, call *CallSession,
 			ControlID: *ctrlID,
 		},
 	}
+
+	savePayload(payload, v)
 
 	var ReplyBladeExecuteDecode ReplyBladeExecute
 
@@ -592,7 +603,7 @@ func (relay *RelaySession) RelayPlayStop(ctx context.Context, call *CallSession,
 }
 
 // RelayRecordAudio TODO DESCRIPTION
-func (relay *RelaySession) RelayRecordAudio(ctx context.Context, call *CallSession, controlID string, rec *RecordParams) error {
+func (relay *RelaySession) RelayRecordAudio(ctx context.Context, call *CallSession, controlID string, rec *RecordParams, payload **json.RawMessage) error {
 	if len(call.CallID) == 0 {
 		Log.Error("no CallID\n")
 
@@ -623,6 +634,8 @@ func (relay *RelaySession) RelayRecordAudio(ctx context.Context, call *CallSessi
 			Record:    record,
 		},
 	}
+
+	savePayload(payload, v)
 
 	call.Lock()
 
@@ -665,7 +678,7 @@ func (relay *RelaySession) RelayRecordAudio(ctx context.Context, call *CallSessi
 }
 
 // RelayRecordAudioStop TODO DESCRIPTION
-func (relay *RelaySession) RelayRecordAudioStop(ctx context.Context, call *CallSession, ctrlID *string) error {
+func (relay *RelaySession) RelayRecordAudioStop(ctx context.Context, call *CallSession, ctrlID *string, payload **json.RawMessage) error {
 	if len(call.CallID) == 0 {
 		Log.Error("no CallID\n")
 
@@ -681,6 +694,8 @@ func (relay *RelaySession) RelayRecordAudioStop(ctx context.Context, call *CallS
 			ControlID: *ctrlID,
 		},
 	}
+
+	savePayload(payload, v)
 
 	var ReplyBladeExecuteDecode ReplyBladeExecute
 
@@ -704,7 +719,7 @@ func (relay *RelaySession) RelayRecordAudioStop(ctx context.Context, call *CallS
 }
 
 // RelayDetectDigit TODO DESCRIPTION
-func (relay *RelaySession) RelayDetectDigit(ctx context.Context, call *CallSession, controlID string, digits string) error {
+func (relay *RelaySession) RelayDetectDigit(ctx context.Context, call *CallSession, controlID string, digits string, payload **json.RawMessage) error {
 	if len(call.CallID) == 0 {
 		Log.Error("no CallID\n")
 
@@ -733,11 +748,11 @@ func (relay *RelaySession) RelayDetectDigit(ctx context.Context, call *CallSessi
 		Log.Debug("controlID was not sent to go routine\n")
 	}
 
-	return relay.RelayDetect(ctx, call, controlID, detect)
+	return relay.RelayDetect(ctx, call, controlID, detect, payload)
 }
 
 // RelayDetectFax TODO DESCRIPTION
-func (relay *RelaySession) RelayDetectFax(ctx context.Context, call *CallSession, controlID string, faxtone string) error {
+func (relay *RelaySession) RelayDetectFax(ctx context.Context, call *CallSession, controlID string, faxtone string, payload **json.RawMessage) error {
 	if len(call.CallID) == 0 {
 		Log.Error("no CallID\n")
 
@@ -766,11 +781,11 @@ func (relay *RelaySession) RelayDetectFax(ctx context.Context, call *CallSession
 		Log.Debug("controlID was not sent to go routine\n")
 	}
 
-	return relay.RelayDetect(ctx, call, controlID, detect)
+	return relay.RelayDetect(ctx, call, controlID, detect, payload)
 }
 
 // RelayDetectMachine TODO DESCRIPTION
-func (relay *RelaySession) RelayDetectMachine(ctx context.Context, call *CallSession, controlID string, det *DetectMachineParams) error {
+func (relay *RelaySession) RelayDetectMachine(ctx context.Context, call *CallSession, controlID string, det *DetectMachineParams, payload **json.RawMessage) error {
 	if len(call.CallID) == 0 {
 		Log.Error("no CallID\n")
 
@@ -802,11 +817,11 @@ func (relay *RelaySession) RelayDetectMachine(ctx context.Context, call *CallSes
 		Log.Debug("controlID was not sent to go routine\n")
 	}
 
-	return relay.RelayDetect(ctx, call, controlID, detect)
+	return relay.RelayDetect(ctx, call, controlID, detect, payload)
 }
 
 // RelayDetect TODO DESCRIPTION
-func (relay *RelaySession) RelayDetect(ctx context.Context, call *CallSession, controlID string, detect DetectStruct) error {
+func (relay *RelaySession) RelayDetect(ctx context.Context, call *CallSession, controlID string, detect DetectStruct, payload **json.RawMessage) error {
 	if len(call.CallID) == 0 {
 		Log.Error("no CallID\n")
 
@@ -832,6 +847,8 @@ func (relay *RelaySession) RelayDetect(ctx context.Context, call *CallSession, c
 		},
 	}
 
+	savePayload(payload, v)
+
 	var ReplyBladeExecuteDecode ReplyBladeExecute
 
 	reply, err := relay.Blade.BladeExecute(ctx, &v, &ReplyBladeExecuteDecode)
@@ -854,7 +871,7 @@ func (relay *RelaySession) RelayDetect(ctx context.Context, call *CallSession, c
 }
 
 // RelayDetectStop TODO DESCRIPTION
-func (relay *RelaySession) RelayDetectStop(ctx context.Context, call *CallSession, ctrlID *string) error {
+func (relay *RelaySession) RelayDetectStop(ctx context.Context, call *CallSession, ctrlID *string, payload **json.RawMessage) error {
 	if len(call.CallID) == 0 {
 		Log.Error("no CallID\n")
 
@@ -870,6 +887,8 @@ func (relay *RelaySession) RelayDetectStop(ctx context.Context, call *CallSessio
 			ControlID: *ctrlID,
 		},
 	}
+
+	savePayload(payload, v)
 
 	var ReplyBladeExecuteDecode ReplyBladeExecute
 
@@ -893,7 +912,7 @@ func (relay *RelaySession) RelayDetectStop(ctx context.Context, call *CallSessio
 }
 
 // RelaySendFax TODO DESCRIPTION
-func (relay *RelaySession) RelaySendFax(ctx context.Context, call *CallSession, ctrlID *string, doc, id, headerInfo string) error {
+func (relay *RelaySession) RelaySendFax(ctx context.Context, call *CallSession, ctrlID *string, fax *FaxParamsInternal, payload **json.RawMessage) error {
 	if len(call.CallID) == 0 {
 		Log.Error("no CallID\n")
 
@@ -907,11 +926,13 @@ func (relay *RelaySession) RelaySendFax(ctx context.Context, call *CallSession, 
 			NodeID:     call.NodeID,
 			CallID:     call.CallID,
 			ControlID:  *ctrlID,
-			Document:   doc,
-			Identity:   id,
-			HeaderInfo: headerInfo,
+			Document:   fax.doc,
+			Identity:   fax.id,
+			HeaderInfo: fax.headerInfo,
 		},
 	}
+
+	savePayload(payload, v)
 
 	select {
 	case call.CallFaxControlID <- *ctrlID:
@@ -943,7 +964,7 @@ func (relay *RelaySession) RelaySendFax(ctx context.Context, call *CallSession, 
 }
 
 // RelayReceiveFax TODO DESCRIPTION
-func (relay *RelaySession) RelayReceiveFax(ctx context.Context, call *CallSession, ctrlID *string) error {
+func (relay *RelaySession) RelayReceiveFax(ctx context.Context, call *CallSession, ctrlID *string, payload **json.RawMessage) error {
 	if len(call.CallID) == 0 {
 		Log.Error("no CallID\n")
 
@@ -959,6 +980,8 @@ func (relay *RelaySession) RelayReceiveFax(ctx context.Context, call *CallSessio
 			ControlID: *ctrlID,
 		},
 	}
+
+	savePayload(payload, v)
 
 	select {
 	case call.CallFaxControlID <- *ctrlID:
@@ -990,7 +1013,7 @@ func (relay *RelaySession) RelayReceiveFax(ctx context.Context, call *CallSessio
 }
 
 // RelaySendFaxStop TODO DESCRIPTION
-func (relay *RelaySession) RelaySendFaxStop(ctx context.Context, call *CallSession, ctrlID *string) error {
+func (relay *RelaySession) RelaySendFaxStop(ctx context.Context, call *CallSession, ctrlID *string, payload **json.RawMessage) error {
 	if len(call.CallID) == 0 {
 		Log.Error("no CallID\n")
 
@@ -1006,6 +1029,8 @@ func (relay *RelaySession) RelaySendFaxStop(ctx context.Context, call *CallSessi
 			ControlID: *ctrlID,
 		},
 	}
+
+	savePayload(payload, v)
 
 	var ReplyBladeExecuteDecode ReplyBladeExecute
 
@@ -1029,7 +1054,7 @@ func (relay *RelaySession) RelaySendFaxStop(ctx context.Context, call *CallSessi
 }
 
 // RelayReceiveFaxStop TODO DESCRIPTION
-func (relay *RelaySession) RelayReceiveFaxStop(ctx context.Context, call *CallSession, ctrlID *string) error {
+func (relay *RelaySession) RelayReceiveFaxStop(ctx context.Context, call *CallSession, ctrlID *string, payload **json.RawMessage) error {
 	if len(call.CallID) == 0 {
 		Log.Error("no CallID\n")
 
@@ -1045,6 +1070,8 @@ func (relay *RelaySession) RelayReceiveFaxStop(ctx context.Context, call *CallSe
 			ControlID: *ctrlID,
 		},
 	}
+
+	savePayload(payload, v)
 
 	var ReplyBladeExecuteDecode ReplyBladeExecute
 
@@ -1068,7 +1095,7 @@ func (relay *RelaySession) RelayReceiveFaxStop(ctx context.Context, call *CallSe
 }
 
 // RelayTapAudio TODO DESCRIPTION
-func (relay *RelaySession) RelayTapAudio(ctx context.Context, call *CallSession, ctrlID, direction string, device *TapDevice) (TapDevice, error) {
+func (relay *RelaySession) RelayTapAudio(ctx context.Context, call *CallSession, ctrlID, direction string, device *TapDevice, payload **json.RawMessage) (TapDevice, error) {
 	tapAudioParams := TapAudioParams{
 		Direction: direction,
 	}
@@ -1079,11 +1106,11 @@ func (relay *RelaySession) RelayTapAudio(ctx context.Context, call *CallSession,
 	}
 
 	// return source Tap device
-	return relay.RelayTap(ctx, call, ctrlID, tap, device)
+	return relay.RelayTap(ctx, call, ctrlID, tap, device, payload)
 }
 
 // RelayTap TODO DESCRIPTION
-func (relay *RelaySession) RelayTap(ctx context.Context, call *CallSession, controlID string, tap TapStruct, device *TapDevice) (TapDevice, error) {
+func (relay *RelaySession) RelayTap(ctx context.Context, call *CallSession, controlID string, tap TapStruct, device *TapDevice, payload **json.RawMessage) (TapDevice, error) {
 	var srcDevice TapDevice
 
 	if len(call.CallID) == 0 {
@@ -1103,6 +1130,8 @@ func (relay *RelaySession) RelayTap(ctx context.Context, call *CallSession, cont
 			Device:    *device,
 		},
 	}
+
+	savePayload(payload, v)
 
 	call.Lock()
 
@@ -1143,7 +1172,7 @@ func (relay *RelaySession) RelayTap(ctx context.Context, call *CallSession, cont
 }
 
 // RelayTapStop TODO DESCRIPTION
-func (relay *RelaySession) RelayTapStop(ctx context.Context, call *CallSession, ctrlID *string) error {
+func (relay *RelaySession) RelayTapStop(ctx context.Context, call *CallSession, ctrlID *string, payload **json.RawMessage) error {
 	if len(call.CallID) == 0 {
 		Log.Error("no CallID\n")
 
@@ -1159,6 +1188,8 @@ func (relay *RelaySession) RelayTapStop(ctx context.Context, call *CallSession, 
 			ControlID: *ctrlID,
 		},
 	}
+
+	savePayload(payload, v)
 
 	var ReplyBladeExecuteDecode ReplyBladeExecute
 
@@ -1182,7 +1213,7 @@ func (relay *RelaySession) RelayTapStop(ctx context.Context, call *CallSession, 
 }
 
 // RelaySendDigits TODO DESCRIPTION
-func (relay *RelaySession) RelaySendDigits(ctx context.Context, call *CallSession, controlID, digits string) error {
+func (relay *RelaySession) RelaySendDigits(ctx context.Context, call *CallSession, controlID, digits string, payload **json.RawMessage) error {
 	if len(call.CallID) == 0 {
 		Log.Error("no CallID\n")
 
@@ -1199,6 +1230,8 @@ func (relay *RelaySession) RelaySendDigits(ctx context.Context, call *CallSessio
 			Digits:    digits,
 		},
 	}
+
+	savePayload(payload, v)
 
 	call.Lock()
 
@@ -1238,7 +1271,7 @@ func (relay *RelaySession) RelaySendDigits(ctx context.Context, call *CallSessio
 }
 
 // RelayPlayAndCollect TODO DESCRIPTION
-func (relay *RelaySession) RelayPlayAndCollect(ctx context.Context, call *CallSession, controlID string, playlist *[]PlayStruct, collect *CollectStruct) error {
+func (relay *RelaySession) RelayPlayAndCollect(ctx context.Context, call *CallSession, controlID string, playlist *[]PlayStruct, collect *CollectStruct, payload **json.RawMessage) error {
 	if len(call.CallID) == 0 {
 		Log.Error("no CallID\n")
 
@@ -1256,6 +1289,8 @@ func (relay *RelaySession) RelayPlayAndCollect(ctx context.Context, call *CallSe
 			Collect:   *collect,
 		},
 	}
+
+	savePayload(payload, v)
 
 	call.Lock()
 
@@ -1301,7 +1336,7 @@ func (relay *RelaySession) RelayPlayAndCollect(ctx context.Context, call *CallSe
 }
 
 // RelayPlayAndCollectVolume TODO DESCRIPTION
-func (relay *RelaySession) RelayPlayAndCollectVolume(ctx context.Context, call *CallSession, ctrlID *string, vol float64) error {
+func (relay *RelaySession) RelayPlayAndCollectVolume(ctx context.Context, call *CallSession, ctrlID *string, vol float64, payload **json.RawMessage) error {
 	if len(call.CallID) == 0 {
 		Log.Error("no CallID\n")
 
@@ -1318,6 +1353,8 @@ func (relay *RelaySession) RelayPlayAndCollectVolume(ctx context.Context, call *
 			Volume:    vol,
 		},
 	}
+
+	savePayload(payload, v)
 
 	var ReplyBladeExecuteDecode ReplyBladeExecute
 
@@ -1341,7 +1378,7 @@ func (relay *RelaySession) RelayPlayAndCollectVolume(ctx context.Context, call *
 }
 
 // RelayPlayAndCollectStop TODO DESCRIPTION
-func (relay *RelaySession) RelayPlayAndCollectStop(ctx context.Context, call *CallSession, ctrlID *string) error {
+func (relay *RelaySession) RelayPlayAndCollectStop(ctx context.Context, call *CallSession, ctrlID *string, payload **json.RawMessage) error {
 	if len(call.CallID) == 0 {
 		Log.Error("no CallID\n")
 
@@ -1357,6 +1394,8 @@ func (relay *RelaySession) RelayPlayAndCollectStop(ctx context.Context, call *Ca
 			ControlID: *ctrlID,
 		},
 	}
+
+	savePayload(payload, v)
 
 	var ReplyBladeExecuteDecode ReplyBladeExecute
 
@@ -1377,4 +1416,26 @@ func (relay *RelaySession) RelayPlayAndCollectStop(ctx context.Context, call *Ca
 	}
 
 	return nil
+}
+
+type placeHolder struct {
+	Params json.RawMessage
+}
+
+func savePayload(payload **json.RawMessage, v interface{}) {
+	if payload != nil {
+		placeholder := new(placeHolder)
+
+		b, err := json.Marshal(v)
+		if err != nil {
+			Log.Error("payload: cannot marshal")
+		}
+
+		err = json.Unmarshal(b, placeholder)
+		if err != nil {
+			Log.Error("payload: cannot unmarshal to RawMessage")
+		}
+
+		*payload = &placeholder.Params
+	}
 }